Add app handlers
