+++ templates/holding.html
@@ -61,7 +61,7 @@
             </span>
             <ul class="p-navigation__links u-align-text--right" role="menu">
               <li class="p-navigation__link" role="menuitem">
-                <a href="/login">Login</a>
+                <a href="/login"><i class="p-icon--user"></i> Login</a>
               </li>
             </ul>
           </nav>
